Move Pods project generation to an Xcode namespace
* Fix Dangerfile Rubocop violations

* Update CHANGELOG.md

* Update docs

* Don't use Type = Type::Type in tests
