+++ .rubocop_todo.yml
@@ -1,12 +1,12 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2016-03-03 16:16:26 -0600 using RuboCop version 0.37.2.
+# on 2016-06-09 08:12:30 -0500 using RuboCop version 0.37.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 4
+# Offense count: 7
 # Cop supports --auto-correct.
 # Configuration parameters: AlignWith, SupportedStyles, AutoCorrect.
 # SupportedStyles: keyword, variable, start_of_line
@@ -23,19 +23,36 @@ Lint/IneffectiveAccessModifier:
     - 'lib/cocoapods/installer/user_project_integrator/target_integrator/xcconfig_integrator.rb'
 
 # Offense count: 1
+# Cop supports --auto-correct.
+# Configuration parameters: AutoCorrect.
+Lint/LiteralInInterpolation:
+  Exclude:
+    - 'Dangerfile'
+
+# Offense count: 1
 Lint/NonLocalExitFromIterator:
   Exclude:
     - 'spec/unit/installer_spec.rb'
+
+# Offense count: 1
+Lint/UnreachableCode:
+  Exclude:
+    - 'Dangerfile'
 
 # Offense count: 40
 Lint/UselessAccessModifier:
   Enabled: false
 
-# Offense count: 1853
+# Offense count: 2035
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes.
 # URISchemes: http, https
 Metrics/LineLength:
   Max: 279
+
+# Offense count: 1
+# Configuration parameters: CountKeywordArgs.
+ParameterLists:
+  Max: 6
 
 # Offense count: 5
 # Cop supports --auto-correct.
@@ -55,20 +72,12 @@ Performance/RedundantBlockCall:
     - 'lib/cocoapods/installer/analyzer/pod_variant_set.rb'
     - 'spec/unit/validator_spec.rb'
 
-# Offense count: 2
-# Cop supports --auto-correct.
-Performance/RedundantMatch:
-  Exclude:
-    - 'spec/**/*'
-    - 'lib/cocoapods/command/lib.rb'
-    - 'lib/cocoapods/external_sources/podspec_source.rb'
-
 # Offense count: 1
 # Cop supports --auto-correct.
 # Configuration parameters: MaxKeyValuePairs.
 Performance/RedundantMerge:
   Exclude:
-    - 'lib/cocoapods/installer/target_installer.rb'
+    - 'lib/cocoapods/installer/xcode/pods_project_generator/target_installer.rb'
 
 # Offense count: 1
 # Cop supports --auto-correct.
@@ -76,7 +85,7 @@ Performance/StringReplacement:
   Exclude:
     - 'spec/functional/command/spec_spec.rb'
 
-# Offense count: 23
+# Offense count: 22
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, SupportedStyles.
 # SupportedStyles: prefer_alias, prefer_alias_method
@@ -106,12 +115,11 @@ Style/AlignParameters:
   Exclude:
     - 'lib/cocoapods/project.rb'
 
-# Offense count: 7
+# Offense count: 6
 # Cop supports --auto-correct.
 # Configuration parameters: SingleLineConditionsOnly.
 Style/ConditionalAssignment:
   Exclude:
-    - 'lib/cocoapods/command/init.rb'
     - 'lib/cocoapods/command/repo/push.rb'
     - 'lib/cocoapods/external_sources/abstract_external_source.rb'
     - 'lib/cocoapods/sandbox/file_accessor.rb'
@@ -137,32 +145,30 @@ Style/IfInsideElse:
 Style/IndentArray:
   EnforcedStyle: consistent
 
-# Offense count: 27
+# Offense count: 24
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, SupportedStyles, IndentationWidth.
 # SupportedStyles: aligned, indented
 Style/MultilineMethodCallIndentation:
   Enabled: false
 
-# Offense count: 1
+# Offense count: 3
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, SupportedStyles, IndentationWidth.
 # SupportedStyles: aligned, indented
 Style/MultilineOperationIndentation:
   Enabled: false
 
-# Offense count: 16
+# Offense count: 12
 # Cop supports --auto-correct.
 Style/MutableConstant:
   Exclude:
     - 'bin/sandbox-pod'
-    - 'lib/cocoapods/command/inter_process_communication.rb'
-    - 'lib/cocoapods/command/lib.rb'
     - 'lib/cocoapods/config.rb'
     - 'lib/cocoapods/generator/copy_resources_script.rb'
     - 'lib/cocoapods/installer.rb'
-    - 'lib/cocoapods/installer/target_installer/pod_target_installer.rb'
     - 'lib/cocoapods/installer/user_project_integrator/target_integrator/xcconfig_integrator.rb'
+    - 'lib/cocoapods/installer/xcode/pods_project_generator/pod_target_installer.rb'
     - 'lib/cocoapods/project.rb'
     - 'spec/unit/installer/user_project_integrator/target_integrator/xcconfig_integrator_spec.rb'
 
@@ -173,16 +179,8 @@ Style/NestedParenthesizedCalls:
 
 # Offense count: 1
 # Cop supports --auto-correct.
-# Configuration parameters: AllowSafeAssignment.
-Style/ParenthesesAroundCondition:
-  Exclude:
-    - 'lib/cocoapods/command/inter_process_communication.rb'
-
-# Offense count: 2
-# Cop supports --auto-correct.
 Style/RedundantParentheses:
   Exclude:
-    - 'lib/cocoapods/sources_manager.rb'
     - 'spec/unit/sandbox_spec.rb'
 
 # Offense count: 2
@@ -191,15 +189,14 @@ Style/RedundantSelf:
   Exclude:
     - 'lib/cocoapods/command.rb'
 
-# Offense count: 16
+# Offense count: 13
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, SupportedStyles, AllowInnerSlashes.
 # SupportedStyles: slashes, percent_r, mixed
 Style/RegexpLiteral:
   Exclude:
-    - 'lib/cocoapods/installer/file_references_installer.rb'
+    - 'lib/cocoapods/installer/xcode/pods_project_generator/file_references_installer.rb'
     - 'lib/cocoapods/sandbox/path_list.rb'
-    - 'lib/cocoapods/sources_manager.rb'
     - 'lib/cocoapods/validator.rb'
     - 'spec/integration.rb'
     - 'spec/unit/external_sources/path_source_spec.rb'
@@ -228,10 +225,8 @@ Style/UnneededInterpolation:
     - 'lib/cocoapods/validator.rb'
     - 'spec/spec_helper/temporary_repos.rb'
 
-# Offense count: 4
+# Offense count: 2
 Style/ZeroLengthPredicate:
   Exclude:
     - 'lib/cocoapods/command/init.rb'
-    - 'lib/cocoapods/command/project.rb'
     - 'lib/cocoapods/installer/podfile_validator.rb'
-    - 'lib/cocoapods/sources_manager.rb'
