+++ lib/cocoapods/installer.rb
@@ -29,9 +29,7 @@ module Pod
   # source control.
   #
   class Installer
-    autoload :AggregateTargetInstaller,   'cocoapods/installer/target_installer/aggregate_target_installer'
     autoload :Analyzer,                   'cocoapods/installer/analyzer'
-    autoload :FileReferencesInstaller,    'cocoapods/installer/file_references_installer'
     autoload :InstallationOptions,        'cocoapods/installer/installation_options'
     autoload :PostInstallHooksContext,    'cocoapods/installer/post_install_hooks_context'
     autoload :PreInstallHooksContext,     'cocoapods/installer/pre_install_hooks_context'
@@ -40,9 +38,8 @@ module Pod
     autoload :PodfileValidator,           'cocoapods/installer/podfile_validator'
     autoload :PodSourceInstaller,         'cocoapods/installer/pod_source_installer'
     autoload :PodSourcePreparer,          'cocoapods/installer/pod_source_preparer'
-    autoload :PodTargetInstaller,         'cocoapods/installer/target_installer/pod_target_installer'
-    autoload :TargetInstaller,            'cocoapods/installer/target_installer'
     autoload :UserProjectIntegrator,      'cocoapods/installer/user_project_integrator'
+    autoload :Xcode,                      'cocoapods/installer/xcode'
 
     include Config::Mixin
     include InstallationOptions::Mixin
@@ -163,15 +160,25 @@ module Pod
       end
     end
 
-    def generate_pods_project
+    #-------------------------------------------------------------------------#
+
+    # @!group Pods Project Generation
+
+    private
+
+    def create_generator
+      Xcode::PodsProjectGenerator.new(aggregate_targets, sandbox, pod_targets, analysis_result, installation_options, config)
+    end
+
+    # Generate the 'Pods/Pods.xcodeproj' project.
+    #
+    def generate_pods_project(generator = create_generator)
       UI.section 'Generating Pods project' do
-        prepare_pods_project
-        install_file_references
-        install_libraries
-        set_target_dependencies
+        generator.generate!
+        @pods_project = generator.project
         run_podfile_post_install_hooks
-        write_pod_project
-        share_development_pod_schemes
+        generator.write
+        generator.share_development_pod_schemes
         write_lockfiles
       end
     end
@@ -551,181 +558,6 @@ module Pod
       end
     end
 
-    # Creates the Pods project from scratch if it doesn't exists.
-    #
-    # @return [void]
-    #
-    # @todo   Clean and modify the project if it exists.
-    #
-    def prepare_pods_project
-      UI.message '- Creating Pods project' do
-        @pods_project = if object_version = aggregate_targets.map(&:user_project).compact.map { |p| p.object_version.to_i }.min
-                          Pod::Project.new(sandbox.project_path, false, object_version)
-                        else
-                          Pod::Project.new(sandbox.project_path)
-        end
-
-        analysis_result.all_user_build_configurations.each do |name, type|
-          @pods_project.add_build_configuration(name, type)
-        end
-
-        pod_names = pod_targets.map(&:pod_name).uniq
-        pod_names.each do |pod_name|
-          local = sandbox.local?(pod_name)
-          path = sandbox.pod_dir(pod_name)
-          was_absolute = sandbox.local_path_was_absolute?(pod_name)
-          @pods_project.add_pod_group(pod_name, path, local, was_absolute)
-        end
-
-        if config.podfile_path
-          @pods_project.add_podfile(config.podfile_path)
-        end
-
-        sandbox.project = @pods_project
-        platforms = aggregate_targets.map(&:platform)
-        osx_deployment_target = platforms.select { |p| p.name == :osx }.map(&:deployment_target).min
-        ios_deployment_target = platforms.select { |p| p.name == :ios }.map(&:deployment_target).min
-        watchos_deployment_target = platforms.select { |p| p.name == :watchos }.map(&:deployment_target).min
-        tvos_deployment_target = platforms.select { |p| p.name == :tvos }.map(&:deployment_target).min
-        @pods_project.build_configurations.each do |build_configuration|
-          build_configuration.build_settings['MACOSX_DEPLOYMENT_TARGET'] = osx_deployment_target.to_s if osx_deployment_target
-          build_configuration.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = ios_deployment_target.to_s if ios_deployment_target
-          build_configuration.build_settings['WATCHOS_DEPLOYMENT_TARGET'] = watchos_deployment_target.to_s if watchos_deployment_target
-          build_configuration.build_settings['TVOS_DEPLOYMENT_TARGET'] = tvos_deployment_target.to_s if tvos_deployment_target
-          build_configuration.build_settings['STRIP_INSTALLED_PRODUCT'] = 'NO'
-          build_configuration.build_settings['CLANG_ENABLE_OBJC_ARC'] = 'YES'
-        end
-      end
-    end
-
-    # Installs the file references in the Pods project. This is done once per
-    # Pod as the same file reference might be shared by multiple aggregate
-    # targets.
-    #
-    # @return [void]
-    #
-    def install_file_references
-      installer = FileReferencesInstaller.new(sandbox, pod_targets, pods_project)
-      installer.install!
-    end
-
-    # Installs the aggregate targets of the Pods projects and generates their
-    # support files.
-    #
-    # @return [void]
-    #
-    def install_libraries
-      UI.message '- Installing targets' do
-        pod_targets.sort_by(&:name).each do |pod_target|
-          target_installer = PodTargetInstaller.new(sandbox, pod_target)
-          target_installer.install!
-        end
-
-        aggregate_targets.sort_by(&:name).each do |target|
-          target_installer = AggregateTargetInstaller.new(sandbox, target)
-          target_installer.install!
-        end
-
-        # TODO: Move and add specs
-        pod_targets.sort_by(&:name).each do |pod_target|
-          pod_target.file_accessors.each do |file_accessor|
-            file_accessor.spec_consumer.frameworks.each do |framework|
-              if pod_target.should_build?
-                pod_target.native_target.add_system_framework(framework)
-              end
-            end
-          end
-        end
-      end
-    end
-
-    # Adds a target dependency for each pod spec to each aggregate target and
-    # links the pod targets among each other.
-    #
-    # @return [void]
-    #
-    def set_target_dependencies
-      frameworks_group = pods_project.frameworks_group
-      aggregate_targets.each do |aggregate_target|
-        is_app_extension = !(aggregate_target.user_targets.map(&:symbol_type) &
-          [:app_extension, :watch_extension, :watch2_extension, :tv_extension]).empty?
-        is_app_extension ||= aggregate_target.user_targets.any? { |ut| ut.common_resolved_build_setting('APPLICATION_EXTENSION_API_ONLY') == 'YES' }
-
-        aggregate_target.pod_targets.each do |pod_target|
-          configure_app_extension_api_only_for_target(aggregate_target) if is_app_extension
-
-          unless pod_target.should_build?
-            pod_target.resource_bundle_targets.each do |resource_bundle_target|
-              aggregate_target.native_target.add_dependency(resource_bundle_target)
-            end
-
-            next
-          end
-
-          aggregate_target.native_target.add_dependency(pod_target.native_target)
-          configure_app_extension_api_only_for_target(pod_target) if is_app_extension
-
-          pod_target.dependent_targets.each do |pod_dependency_target|
-            next unless pod_dependency_target.should_build?
-            pod_target.native_target.add_dependency(pod_dependency_target.native_target)
-            configure_app_extension_api_only_for_target(pod_dependency_target) if is_app_extension
-
-            if pod_target.requires_frameworks?
-              product_ref = frameworks_group.files.find { |f| f.path == pod_dependency_target.product_name } ||
-                frameworks_group.new_product_ref_for_target(pod_dependency_target.product_basename, pod_dependency_target.product_type)
-              pod_target.native_target.frameworks_build_phase.add_file_reference(product_ref, true)
-            end
-          end
-        end
-      end
-    end
-
-    # Writes the Pods project to the disk.
-    #
-    # @return [void]
-    #
-    def write_pod_project
-      UI.message "- Writing Xcode project file to #{UI.path sandbox.project_path}" do
-        pods_project.pods.remove_from_project if pods_project.pods.empty?
-        pods_project.development_pods.remove_from_project if pods_project.development_pods.empty?
-        pods_project.sort(:groups_position => :below)
-        if installation_options.deterministic_uuids?
-          UI.message('- Generating deterministic UUIDs') { pods_project.predictabilize_uuids }
-        end
-        pods_project.recreate_user_schemes(false)
-        pods_project.save
-      end
-    end
-
-    # Shares schemes of development Pods.
-    #
-    # @return [void]
-    #
-    def share_development_pod_schemes
-      development_pod_targets.select(&:should_build?).each do |pod_target|
-        next unless share_scheme_for_development_pod?(pod_target.pod_name)
-        Xcodeproj::XCScheme.share_scheme(pods_project.path, pod_target.label)
-      end
-    end
-
-    # @param  [String] pod The root name of the development pod.
-    #
-    # @return [Bool] whether the scheme for the given development pod should be
-    #         shared.
-    #
-    def share_scheme_for_development_pod?(pod)
-      case dev_pods_to_share = installation_options.share_schemes_for_development_pods
-      when TrueClass, FalseClass, NilClass
-        dev_pods_to_share
-      when Array
-        dev_pods_to_share.any? { |dev_pod| dev_pod === pod } # rubocop:disable Style/CaseEquality
-      else
-        raise Informative, 'Unable to handle share_schemes_for_development_pods ' \
-          "being set to #{dev_pods_to_share.inspect} -- please set it to true, " \
-          'false, or an array of pods to share schemes for.'
-      end
-    end
-
     # Writes the Podfile and the lock files.
     #
     # @todo   Pass the checkout options to the Lockfile.
@@ -857,15 +689,6 @@ module Pod
     #
     def sandbox_state
       analysis_result.sandbox_state
-    end
-
-    # Sets the APPLICATION_EXTENSION_API_ONLY build setting to YES for all
-    # configurations of the given target
-    #
-    def configure_app_extension_api_only_for_target(target)
-      target.native_target.build_configurations.each do |config|
-        config.build_settings['APPLICATION_EXTENSION_API_ONLY'] = 'YES'
-      end
     end
 
     #-------------------------------------------------------------------------#
