+++ lib/cocoapods/installer/xcode.rb
@@ -1,0 +1,7 @@
+module Pod
+  class Installer
+    class Xcode
+      autoload :PodsProjectGenerator, 'cocoapods/installer/xcode/pods_project_generator'
+    end
+  end
+end
